Fix circular reference crash (#2)
* Fix circular reference crash in to_json

The crash was because the to_json helper function in v8context.cc
did not check to ensure that the JS stringify call succeeded
before calling ToString() on the possibly-empty result.

Now in the to_json helper function in v8context.cc we return the
empty string and re-throw exceptions that we catch. Callers can
check the return value for the empty string (an invalid JSON value)
and deal with the re-thrown error as we do elsewhere in the file.

The golang ToJSON function should really return an error, since
it is not always possible to convert a v8 value to JSON, as in
the circular reference case. So updated it and its callers
accordingly.